BOT/crypto-alpha-bot/requirements.txt
twikit>=2.2.0
anthropic>=0.40.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio
